backend/pyproject.toml
[project]
name = "llm_data_lab_backend"
version = "0.1.0"
description = "Backend service for the LLM Data Lab platform"
requires-python = ">=3.10"
authors = [{name = "LLM Data Lab Team"}]
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.6.1",
    "email-validator>=2.1.0",
    "pydantic-settings>=2.2.1",
    "pandas>=2.2.0",
    "numpy>=1.26.3",
    "matplotlib>=3.8.2",
    "seaborn>=0.13.2",
    "plotly>=5.18.0",
    "scipy>=1.12.0",
    "statsmodels>=0.14.1",
    "scikit-learn>=1.4.0",
    "lifelines>=0.28.0",
    "pymc>=5.10.0",
    "arviz>=0.17.0",
    "nltk>=3.8.1",
    "spacy>=3.7.2",
    "prophet>=1.1.5",
    "shap>=0.44.1",
    "python-multipart>=0.0.9",
    "openpyxl>=3.1.2",
    "sqlalchemy>=2.0.25",
    "databases[sqlite]>=0.7.0",
    "httpx>=0.26.0",
    "jinja2>=3.1.3",
    "python-dotenv>=1.0.1",
    "cryptography>=42.0.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0"
]

[project.optional-dependencies]
dev = [
    "ruff>=0.2.1",
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3"
]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "UP", "B", "SIM", "ASYNC"]
ignore = ["B008"]

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*"]
